+++ CHANGELOG.md
@@ -78,6 +78,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add deposit single sided with exact amount endpoint
+- Fix `findSwappableMinMaxBinId` returned invalid min/max bin id under some edge case
+- Fix `derivePosition` using invalid seed
 
 ## @meteora-ag/dlmm [1.0.31] - PR #59